moved structure sampler class so it can be imported by other projects
